test: fix all the failing state tests
Fix all the tests that are failing due to a missing charm document.

+++ domain/application/state/config_test.go
@@ -173,8 +173,8 @@ func (s *configSuite) TestEncodeConfigType(c *gc.C) {
 }
 
 func (s *configSuite) TestEncodeConfigDefaultValue(c *gc.C) {
-	_, err := encodeConfigDefaultValue(int64(0))
-	c.Assert(err, gc.ErrorMatches, `unknown config default value type int64`)
+	_, err := encodeConfigDefaultValue(int32(0))
+	c.Assert(err, gc.ErrorMatches, `unknown config default value type int32`)
 }
 
 var configTypeTestCases = [...]struct {
